Finish initial ECS implementation
ECS is split into Entities, Components, and Systems.

Systems manage all of the Components relevant to their Update or Draw
cycle. Each Component is registered to a "Parent" Entity. Entities are
little more than an ID, providing helper methods to interact with the
relevant Components. These helper methods are expensive to call, but the
use case would be that on a BehaviorComponent initialization, the
necessary Components would be retrieved via these slow helper methods.

Systems manage the lifetimes of Components. When an Entity is deleted,
all of the Components are queued up for deletion. On the next Update
cycle, each System will delete whatever components it has queued up for
deletion.

When the Application closes down, all of the Entities are unregistered
and the Components are immediately deleted.

Systems are generalized to accept a list of Types. Currently, an Entity
must have an instance of each Type to be registered to a System.
Whenever a Component is added or removed from an Entity or a System is
created, relevant Entities are checked if they should be registered to
the Application Systems.

An Example of a CustomSystem and a custom Behavior are present in the
BasicSteelGame project.